Update in default title and default body
When explain maker was born the title and body were top fields of DynamoDB table records.
Due to a limitation in the value of DynamoDB table top fields, the title and body needed
to be non empty, and were set to be a single dash. We no longer have this limitation.